refactor: clean js files for frontend and clean auth files
